requirements.txt
toml>=0.10.0
neo4j>=1.7.6
numpy>=1.17.4
scikit-learn>=0.23.2
scikit-surprise>=1.1.1
scipy>=1.5.2